+++ src/main/java/com/reason/lang/ocaml/OclParser.java
@@ -620,8 +620,8 @@ public class OclParser extends CommonPsiParser {
                 } else {
                     popEndUntil(myTypes.C_TRY_EXPR);
                 }
-            } else if (strictlyInAny(myTypes.C_LET_DECLARATION, myTypes.C_MODULE_DECLARATION, myTypes.C_PATTERN_MATCH_BODY)) {
-                boolean isStart = isFound(myTypes.C_LET_DECLARATION) || isFound(myTypes.C_MODULE_DECLARATION);
+            } else if (strictlyInAny(myTypes.C_LET_DECLARATION, myTypes.C_MODULE_DECLARATION, myTypes.C_PATTERN_MATCH_BODY, myTypes.C_OPEN/*local open*/)) {
+                boolean isStart = isFound(myTypes.C_LET_DECLARATION) || isFound(myTypes.C_MODULE_DECLARATION) || isFound(myTypes.C_OPEN);
                 popEndUntilFoundIndex();
                 if (isStart) {
                     popEnd();